# public/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BordApp API",
    "description": "Документація до API бордової дошки",
    "contact": {
      "email": "support@bord-app.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://board.filkx.com",
      "description": "Основний сервер"
    }
  ],
  "paths": {
...
    "/cabinet/adverts": {
      "get": {
        "tags": ["Cabinet Adverts"],
        "summary": "Отримати список оголошень користувача",
        "description": "Повертає пагінований список оголошень, створених поточним користувачем.",
        "operationId": "58bffc737f6867ea061941363fcea815",
        "responses": {
          "200": {
            "description": "Список оголошень",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdvertListResource"
                      }
                    },
                    "links": {
                      "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthenticated."
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cabinet Adverts"],
        "summary": "Створити нове оголошення",
        "description": "Створює нове оголошення для поточного користувача.",
        "operationId": "15d7ba68372d3ac53cb706feb6ea8a84",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Оголошення створено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvertDetailResource"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthenticated."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The given data was invalid."
                    },
                    "errors": {
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/cabinet/adverts/{advert}/attributes": {
      "put": {
        "tags": ["Cabinet Adverts"],
        "summary": "Оновити атрибути оголошення",
        "description": "Оновлює атрибути існуючого оголошення користувача.",
        "operationId": "85027be7ef1d2867e9e5d26346e180bd",
        "parameters": [
          {
            "name": "advert",
            "in": "path",
            "description": "ID оголошення",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 123
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Атрибути оголошення оновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvertDetailResource"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthenticated."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Доступ заборонено",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "This action is unauthorized."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Оголошення не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Advert not found."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The given data was invalid."
                    },
                    "errors": {
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/cabinet/adverts/{advert}/photos": {
      "put": {
        "tags": ["Cabinet Adverts"],
        "summary": "Додати фото до оголошення",
        "description": "Додає фото до існуючого оголошення користувача.",
        "operationId": "80a0abf1fd26b4734e13e0518ad47fed",
        "parameters": [
          {
            "name": "advert",
            "in": "path",
            "description": "ID оголошення",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 123
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotosRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Фото додано до оголошення",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvertDetailResource"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизований доступ",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthenticated."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Доступ заборонено",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "This action is unauthorized."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Оголошення не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Advert not found."
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Помилка валідації",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The given data was invalid."
                    },
                    "errors": {
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdvertListResource": {
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user": {
            "properties": {
              "name": {
                "type": "string",
                "example": "Іван"
              },
              "phone": {
                "type": "string",
                "example": "+380991234567"
              }
            },
            "type": "object"
          },
          "category": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 4
              },
              "name": {
                "type": "string",
                "example": "Авто"
              }
            },
            "type": "object"
          },
          "region": {
            "properties": {
              "id": {
                "type": "integer",
                "example": 22973
              },
              "name": {
                "type": "string",
                "example": "Київ"
              }
            },
            "type": "object"
          },
          "title": {
            "type": "string",
            "example": "Продаю BMW E39"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 15000
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T12:00:00Z"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}